Move `DialogueHandlerExt::add_dialogue` to `HandlerExt`
